1) Changed ComparativeGenomeAssembler to use temporary files and paths instead of hardcoded filenames. 2) Updates for Padena from Nigel 3) Unit test fixes 4) Added Try methods to ProteinTranslation 5) Dropped FastAParser buffer size down to avoid OutOfMemory exceptions on 32-bit machines 6) Added echo to console for application logging. 7) Fixed DeltaAlignmentParser to not keep sequence file open after Dispose and removed finalizer.

+++ bio/Source/Framework/Bio/Assembly/Padena/ILowCoverageContigPurger.cs
@@ -14,6 +14,6 @@ namespace Bio.Algorithms.Assembly.Padena
         /// <param name="deBruijnGraph">DeBruijn Graph.</param>
         /// <param name="coverageThresholdForContigs">Coverage Threshold for contigs.</param>
         /// <returns>Number of nodes removed.</returns>
-        int RemoveLowCoverageContigs(DeBruijnGraph deBruijnGraph, double coverageThresholdForContigs);
+        long RemoveLowCoverageContigs(DeBruijnGraph deBruijnGraph, double coverageThresholdForContigs);
     }
 }
